+++ test/model/bfm/cclo_bfm.cpp
@@ -18,7 +18,7 @@
 #include "cclo_bfm.h"
 #include "zmq_client.h"
 #include <stdexcept>
-#include "constants.hpp"
+#include "accl/constants.hpp"
 #include <iostream>
 #include <mutex>
